Revert "Feat/testes entidades generos"
